requirements.txt
Flask
SpeechRecognition
pyttsx3
python-dotenv
